Payment Method Selector v2 (#274)
* Updated the JPSession object with the new endpoint

* Reworked the iDEAL Service

* Added the iDEAL bank view model

* Ran ./clang-format.sh

* Fixed an issue with renaming endpoint to baseURL

* Added the checkmark icon logic

* Ran ./clang-format.sh

* Added constants to JPSectionView

* Updated the card header for iDEAL banks

* Ran ./clang-format.sh

* [WIP] Added iDEAL calls

* [WIP] Added pooling logic

* Re-added ExampleAppCredentials

* Localized strings and added iDEAL checks

* Added missing Site ID error and constants

* Added JPTheme to the JPUIConfiguration object

* Added ability to set custom UIColors in the JPTheme object

* Customized card payment method

* Added customization for the header view

* Fully customized the Payment Method Screen

* Added Transaction View customization

* Customized iDEAL flow

* Customized Scan Card screen

* Added button customization

* Added toggle for showing the amount

* Added ability to set custom back button image for the merchant

* [WIP] Added configuration for JPInputField

* [WIP] Even more JPInputField customization

* Ran ./clang-format.sh

* Re-added missing UITextField extension

* Fixed documentation issues and fixed property naming

* Added missing documentation

* Fixed PR comments

* Fixed PR comments

* Fixed PR comments

* Fixed PR comments

* Fixed PR comments

* Added a better JPSectionView implementation

* Fixed PR comments

* Fixed PR comments part 2

* Documented JPSection

+++ Source/Modules/PaymentMethods/Presenter/JPPaymentMethodsPresenter.h
@@ -86,7 +86,7 @@
  *
  * @param index - the index of the newly selected payment method
  */
-- (void)changePaymentMethodToIndex:(int)index;
+- (void)changePaymentMethodToIndex:(NSUInteger)index;
 
 /**
  * A method that sets the latest added card as selected
